feat: [SRM-13376]: Adding API impl for improved verify step UX. (#42020)
* [SRM-13376]: Adding complete API impl

* [SRM-13376]: Addressing review comments. Changes to include interfaces in swagger documentation.

+++ srm-service/modules/cv-nextgen-service/contracts/commons/src/main/java/io/harness/cvng/cdng/beans/v2/VerificationSpec.java
@@ -18,6 +18,6 @@ import lombok.experimental.SuperBuilder;
 public class VerificationSpec {
   VerificationJobType analysisType;
   Sensitivity sensitivity;
-  int durationInMinutes;
+  long durationInMinutes;
   Boolean isFailOnNoAnalysis;
 }
